Do not grow pbuf zone at runtime
When we grow the pbuf zone, we must allocate kmem to back the
new pbuf.  This allocation may require the kernel to grow its
pmap.  Growing the kernel pmap is a bit fragile, and can panic
if we are totally out of memory.

Unfortunately, the point at which we grow the pbuf zone is when
the disk queues are backed up and the system is rather sick.  Eg,
the point at which we're most likely to run low on memory.

The simplest solution is to preallocate the zone, and never grow
it.  The original zone max was a rather arbitrary choice, and
never re-examined.  To pick a new, lower zone max, I used a
a 100g / 100% TLS box and picked a value that led to  just a
handful of sleeps.